// src/modules-features/admin/ModuleImprovementPlan/interfaces/ICriteriaInfoViewModel.ts
export interface ICriteriaInfoViewModel {
  id?: number;
  programName?: string;
  evaluationPeriod?: string;
  standardCode?: string;
  criteriaCode?: string;
  criteriaName?: string;
  externalResult?: string;
  externalDetail?: string;
  midtermReportDate?: Date;
  midtermReportCompleteDate?: Date;
  staffCode?: string;
  staffName?: string;
}